Refactor cooking
