+++ ChopChop/Models/Cooking/SessionRecipeStepNode.swift
@@ -1,17 +1,16 @@
-import Combine
 import CoreGraphics
 import Foundation
 
-final class SessionRecipeStepNode: Node, ObservableObject {
+final class SessionRecipeStepNode: Node {
     let id = UUID()
     var label: SessionRecipeStep
     var position: CGPoint?
-    @Published var isCompletable = false
-    @Published var isCompleted = false
 
-    init(_ node: RecipeStepNode, actionTimeTracker: ActionTimeTracker, position: CGPoint? = nil) {
-        self.label = SessionRecipeStep(step: node.label, actionTimeTracker: actionTimeTracker)
-        self.position = position
+    var isCompletable = true
+    var isCompleted = false
+
+    init(node: RecipeStepNode) {
+        self.label = SessionRecipeStep(step: node.label)
     }
 }
