+++ ChopChop/ViewModels/Graph/SessionGraphViewModel.swift
@@ -2,7 +2,6 @@ import SwiftUI
 
 final class SessionGraphViewModel: ObservableObject {
     @Published var portalPosition = CGVector.zero
-    @Published var showTimerPanel = false
 
     var graph: SessionRecipeStepGraph
