+++ ChopChop/Views/Cooking/CountdownTimerView.swift
@@ -2,76 +2,122 @@ import SwiftUI
 
 struct CountdownTimerView: View {
     @ObservedObject var viewModel: CountdownTimerViewModel
+    @State var opacity = true
 
     var body: some View {
-        HStack {
-            Text("Timer")
-            timeDisplay
-            if !viewModel.countdownTimer.isStart {
-                VStack {
-                    plusTimeButton
-                    minusTimeButton
-                }
-                startButton
-            } else {
-                pauseButton
-                restartButton
+        HStack(spacing: 0) {
+            switch viewModel.status {
+            case .stopped:
+                stoppedView
+            case .running:
+                runningView
+            case .paused:
+                pausedView
+            case .ended:
+                endedView
             }
         }
-        .disabled(viewModel.isDisabled)
-        .frame(width: 220, height: 60, alignment: .center)
+        .padding([.leading, .trailing], 8)
+        .frame(width: 185)
         .foregroundColor(.white)
-        .background(Color.blue)
+        .background(background)
         .clipShape(Capsule())
     }
 
-    var timeDisplay: some View {
-        Text(viewModel.displayTime)
-            .onReceive(viewModel.countdownTimer.timer) { _ in
-                viewModel.countdown()
+    @ViewBuilder
+    var stoppedView: some View {
+        Text(viewModel.timeRemaining)
+            .font(Font.body.monospacedDigit())
+            .padding()
+            .onTapGesture(perform: viewModel.timer.start)
+        Button(action: viewModel.timer.start) {
+            Image(systemName: "play.fill")
+        }
+        .padding(EdgeInsets(top: 16, leading: 8, bottom: 16, trailing: 8))
+        Button(action: viewModel.timer.reset) {
+            Image(systemName: "arrow.clockwise")
+        }
+        .padding(EdgeInsets(top: 16, leading: 8, bottom: 16, trailing: 8))
+    }
+
+    @ViewBuilder
+    var runningView: some View {
+        Text(viewModel.timeRemaining)
+            .font(Font.body.monospacedDigit())
+            .padding()
+            .onTapGesture(perform: viewModel.timer.pause)
+        Spacer()
+        Button(action: viewModel.timer.pause) {
+            Image(systemName: "pause.fill")
+        }
+        .padding(EdgeInsets(top: 16, leading: 8, bottom: 16, trailing: 8))
+    }
+
+    @ViewBuilder
+    var pausedView: some View {
+        Text(viewModel.timeRemaining)
+            .font(Font.body.monospacedDigit())
+            .padding()
+            .opacity(opacity ? 1 : 0.4)
+            .animation(Animation.linear(duration: .leastNonzeroMagnitude).delay(0.5).repeatForever())
+            .onAppear {
+                opacity.toggle()
             }
-            .foregroundColor(viewModel.countdownTimer.remainingTime == 0 ? .red : .white)
+            .onTapGesture(perform: viewModel.timer.resume)
+        Button(action: viewModel.timer.resume) {
+            Image(systemName: "play.fill")
+        }
+        .padding(EdgeInsets(top: 16, leading: 8, bottom: 16, trailing: 8))
+        Button(action: viewModel.timer.reset) {
+            Image(systemName: "arrow.clockwise")
+        }
+        .padding(EdgeInsets(top: 16, leading: 8, bottom: 16, trailing: 8))
     }
 
-    var plusTimeButton: some View {
-        Button("+") {
-            viewModel.increaseTime()
+    @ViewBuilder
+    var endedView: some View {
+        Text(viewModel.timeRemaining)
+            .font(Font.body.monospacedDigit())
+            .padding()
+            .opacity(opacity ? 1 : 0.4)
+            .animation(Animation.linear(duration: .leastNonzeroMagnitude).delay(0.5).repeatForever())
+            .onAppear {
+                opacity.toggle()
+            }
+            .onTapGesture(perform: viewModel.timer.reset)
+        Spacer()
+        Button(action: viewModel.timer.reset) {
+            Image(systemName: "stop.fill")
         }
-        .disabled(viewModel.isDisabled)
+        .padding(EdgeInsets(top: 16, leading: 8, bottom: 16, trailing: 8))
     }
 
-    var minusTimeButton: some View {
-        Button("-") {
-            viewModel.decreaseTime()
+    @ViewBuilder
+    var background: some View {
+        switch viewModel.status {
+        case .running, .paused:
+            ZStack(alignment: .leading) {
+                Rectangle()
+                    .fill(Color.accentColor)
+                    .overlay(
+                        Rectangle()
+                            .fill(Color(UIColor.systemBackground).opacity(0.2))
+                    )
+                Rectangle()
+                    .fill(Color.accentColor)
+                    .frame(width: 185 * CGFloat(viewModel.timer.timeRemaining / viewModel.timer.duration))
+                    .animation(.easeInOut)
+            }
+        case .stopped:
+            Color.accentColor
+        case .ended:
+            Color.red
         }
-        .disabled(viewModel.isDisabled)
-    }
-
-    var startButton: some View {
-        Button(action: viewModel.start) {
-            Image(systemName: "hourglass.bottomhalf.fill")
-        }
-        .disabled(viewModel.isDisabled)
-    }
-
-    var pauseButton: some View {
-        Button(action: viewModel.pauseResume) {
-            Image(systemName: "playpause")
-        }
-        .disabled(viewModel.isDisabled)
-    }
-
-    var restartButton: some View {
-        Button(action: viewModel.restart) {
-            Image(systemName: "arrow.counterclockwise")
-        }
-        .disabled(viewModel.isDisabled)
     }
 }
 
 struct CountdownTimerView_Previews: PreviewProvider {
     static var previews: some View {
-        // swiftlint:disable force_try
-        CountdownTimerView(viewModel: CountdownTimerViewModel(countdownTimer: try! CountdownTimer(time: 60)))
+        CountdownTimerView(viewModel: CountdownTimerViewModel(timer: CountdownTimer(duration: 10)))
     }
 }
